doubly_linkedlist.c
This is implementation of doubly linked list in C language. I have implemented six basic operations on the linked list. They are Inserting an element at the beginning, inserting an element at  a position entered by the user, inserting an element at the end of the list, deleting an element at the beginning of the list, deleting an element from a position entered by the user, deleting an element from the end of the list. In addition to this i have also implemented the display function to verify whether the elements are being added to or deleted from  their correct positions. It is a menu driven program. Any suggestions are most welcomed.